moved queueProcessor to service directory
